reorganize bitcoin-da to pass check-features
